+++ contracts/sns/SlotLib2.sol
@@ -4,12 +4,12 @@ import "./SlotMachineStorage.sol";
 
 library SlotLib2 {
 
-    event slotMachineCreated(address _provider, uint _decider, uint _minBet, uint _maxBet, uint _maxPrize, uint _totalnum, address _slotaddr);
+    event slotMachineCreated(address _provider, uint16 _decider, uint _minBet, uint _maxBet, uint16 _maxPrize, uint _totalnum, address _slotaddr);
     event slotMachineRemoved(address _provider, address _slotaddr, uint _totalnum);
 
 
     //create new slotmachine
-    function createSlotMachine (address _slotmachineStorage, address _provider,  uint _decider, uint _minBet, uint _maxBet, uint _maxPrize) returns (address) {
+    function createSlotMachine (address _slotmachineStorage, address _provider,  uint16 _decider, uint _minBet, uint _maxBet, uint16 _maxPrize) returns (address) {
         address newslot = address(SlotMachineStorage(_slotmachineStorage).createSlotMachine(_provider, _decider, _minBet, _maxBet, _maxPrize));
         slotMachineCreated(_provider, _decider, _minBet, _maxBet, _maxPrize, SlotMachineStorage(_slotmachineStorage).getNumofSlotMachine(_provider),newslot);
         return newslot;
